Дорабатывает фильтровку фотографий - раскидывает по модулям и убирает лишнее
